application/backend_flask/requirements.txt
wheel
requests
argon2-cffi
flask
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
sqlalchemy
mysqlclient
openpyxl
gunicorn
Authlib
pandas